+++ CHANGELOG.md
@@ -4,7 +4,7 @@ Version numbers follow [Semantic Versioning](https://semver.org/).
 ## Unversioned
 
 - Breaking: Updated `metrics` to version 0.17.
-- Breaking: Implement user login fetching via the API when using `RefreshingLoginCredentials`.
+- Breaking: Implement user login fetching via the API when using `RefreshingLoginCredentials`. (#144)
 - Minor: Implement `Clone` for `RefreshingLoginCredentials`
 
 ## v3.0.1