+++ CHANGELOG.md
@@ -4,7 +4,7 @@ Version numbers follow [Semantic Versioning](https://semver.org/).
 ## Unversioned
 
 - Breaking: Updated `metrics` to version 0.17.
-- Minor: Implement `Clone` for `RefreshingLoginCredentials`
+- Minor: Implement `Clone` for `RefreshingLoginCredentials` (#143)
 
 ## v3.0.1