+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Version numbers follow [Semantic Versioning](https://semver.org/).
 - Breaking: Updated `metrics` to version 0.17.
 - Breaking: Implement user login fetching via the API when using `RefreshingLoginCredentials`. (#144)
 - Minor: Implement `Clone` for `RefreshingLoginCredentials` (#143)
+- Minor: Added `tokio-tracing` feature that adds a name to tasks spawned by this library. (#145)
 
 ## v3.0.1