+++ CHANGELOG.md
@@ -10,6 +10,13 @@ Version numbers follow [Semantic Versioning](https://semver.org/).
 - Breaking: Renamed the `refreshing-token` feature to `refreshing-token-native-tls` to reflect the fact that it pulls in the OS's native TLS library (e.g. OpenSSL/Schannel). Added the `refreshing-token-rustls-native-roots` and `refreshing-token-rustls-webpki-roots` feature flags to complement the other parts of the library where you can choose between the three options. (#153)
 - Minor: Implement `Clone` for `RefreshingLoginCredentials` (#143)
 - Minor: Added feature flag `transport-ws-rustls-native-roots` to allow websocket connections powered by rustls using the OS-native root certificates. (#146)
+- Minor: Updated dependencies:
+  - `async-tungstenite` `0.13` -> `0.17`
+  - `metrics` `0.16` -> `0.18`
+  - `rustls-native-certs` `0.5` -> `0.6`
+  - `tokio-rustls` `0.22` -> `0.23`
+  - `tokio-util` `0.6` -> `0.7`
+  - `webpki-roots` `0.21` -> `0.22`
 
 ## v3.0.1